backend/requirements.txt
numpy
pandas
tqdm
torch
torchvision
matplotlib
seaborn
openai
langchain
streamlit
python-dotenv
requests
beautifulsoup4
googlesearch-python
ipykernel
newspaper3k
lxml_html_clean
langchain_groq
datetime
ffmpeg-python
assemblyai
pillow
opencv-python
scikit-learn
facenet-pytorch 